Rename ShelterTypes to clarify their purpose
ShelterType in accessibility properties -> ShelterWidthType
HslShelterType in shelter properties -> ShelterType
