+++ src/main/java/org/rutebanken/tiamat/model/ShelterEquipment_VersionStructure.java
@@ -21,7 +21,7 @@ import javax.persistence.MappedSuperclass;
 import javax.persistence.Transient;
 import java.math.BigDecimal;
 import org.rutebanken.tiamat.model.hsl.ElectricityTypeEnumeration;
-import org.rutebanken.tiamat.model.hsl.HslShelterTypeEnumeration;
+import org.rutebanken.tiamat.model.hsl.ShelterTypeEnumeration;
 import org.rutebanken.tiamat.model.hsl.ShelterConditionEnumeration;
 
 @MappedSuperclass
@@ -33,7 +33,7 @@ public class ShelterEquipment_VersionStructure
     protected BigDecimal distanceFromNearestKerb;
 
     @Enumerated(EnumType.STRING)
-    protected HslShelterTypeEnumeration shelterType;
+    protected ShelterTypeEnumeration shelterType;
 
     @Enumerated(EnumType.STRING)
     protected ElectricityTypeEnumeration shelterElectricity;
@@ -73,11 +73,11 @@ public class ShelterEquipment_VersionStructure
         this.distanceFromNearestKerb = value;
     }
 
-    public HslShelterTypeEnumeration getShelterType() {
+    public ShelterTypeEnumeration getShelterType() {
         return shelterType;
     }
 
-    public void setShelterType(HslShelterTypeEnumeration shelterType) {
+    public void setShelterType(ShelterTypeEnumeration shelterType) {
         this.shelterType = shelterType;
     }
