+++ src/test/java/org/rutebanken/tiamat/rest/graphql/GraphQLResourceStopPlaceIntegrationTest.java
@@ -33,12 +33,12 @@ import org.rutebanken.tiamat.model.GeneralSign;
 import org.rutebanken.tiamat.model.hsl.ElectricityTypeEnumeration;
 import org.rutebanken.tiamat.model.hsl.GuidanceTypeEnumeration;
 import org.rutebanken.tiamat.model.hsl.HslAccessibilityProperties;
-import org.rutebanken.tiamat.model.hsl.HslShelterTypeEnumeration;
+import org.rutebanken.tiamat.model.hsl.ShelterTypeEnumeration;
 import org.rutebanken.tiamat.model.hsl.HslStopTypeEnumeration;
 import org.rutebanken.tiamat.model.hsl.MapTypeEnumeration;
 import org.rutebanken.tiamat.model.hsl.PedestrianCrossingRampTypeEnumeration;
 import org.rutebanken.tiamat.model.hsl.ShelterConditionEnumeration;
-import org.rutebanken.tiamat.model.hsl.ShelterTypeEnumeration;
+import org.rutebanken.tiamat.model.hsl.ShelterWidthTypeEnumeration;
 import org.rutebanken.tiamat.model.InterchangeWeightingEnumeration;
 import org.rutebanken.tiamat.model.LimitationStatusEnumeration;
 import org.rutebanken.tiamat.model.NameTypeEnumeration;
@@ -2297,7 +2297,7 @@ public class GraphQLResourceStopPlaceIntegrationTest extends AbstractGraphQLReso
             .body("stopAreaSurroundingsAccessible", equalTo(true))
             .body("curvedStop", equalTo(true))
             .body("stopType", equalTo(HslStopTypeEnumeration.PULL_OUT.value()))
-            .body("shelterType", equalTo(ShelterTypeEnumeration.WIDE.value()))
+            .body("shelterType", equalTo(ShelterWidthTypeEnumeration.WIDE.value()))
             .body("guidanceType", equalTo(GuidanceTypeEnumeration.BRAILLE.value()))
             .body("mapType", equalTo(MapTypeEnumeration.TACTILE.value()))
             .body("pedestrianCrossingRampType", equalTo(PedestrianCrossingRampTypeEnumeration.RK4_LR.value()));
@@ -2425,7 +2425,7 @@ public class GraphQLResourceStopPlaceIntegrationTest extends AbstractGraphQLReso
             .body("stopAreaSurroundingsAccessible", equalTo(true))
             .body("curvedStop", equalTo(true))
             .body("stopType", equalTo(HslStopTypeEnumeration.BUS_BULB.value()))
-            .body("shelterType", equalTo(ShelterTypeEnumeration.NARROW.value()))
+            .body("shelterType", equalTo(ShelterWidthTypeEnumeration.NARROW.value()))
             .body("guidanceType", equalTo(GuidanceTypeEnumeration.NONE.value()))
             .body("mapType", equalTo(MapTypeEnumeration.OTHER.value()))
             .body("pedestrianCrossingRampType", equalTo(PedestrianCrossingRampTypeEnumeration.LR.value()));
@@ -2679,7 +2679,7 @@ public class GraphQLResourceStopPlaceIntegrationTest extends AbstractGraphQLReso
                 .rootPath("data.stopPlace[0].placeEquipments.shelterEquipment[0]")
                 .body(notNullValue())
                 .body("enclosed", equalTo(true))
-                .body("shelterType", equalTo(HslShelterTypeEnumeration.STEEL.value()))
+                .body("shelterType", equalTo(ShelterTypeEnumeration.STEEL.value()))
                 .body("shelterElectricity", equalTo(ElectricityTypeEnumeration.CONTINUOUS.value()))
                 .body("shelterLighting", equalTo(true))
                 .body("shelterCondition", equalTo(ShelterConditionEnumeration.MEDIOCRE.value()))
@@ -2693,7 +2693,7 @@ public class GraphQLResourceStopPlaceIntegrationTest extends AbstractGraphQLReso
                 .rootPath("data.stopPlace[0].placeEquipments.shelterEquipment[1]")
                 .body(notNullValue())
                 .body("enclosed", equalTo(false))
-                .body("shelterType", equalTo(HslShelterTypeEnumeration.GLASS.value()))
+                .body("shelterType", equalTo(ShelterTypeEnumeration.GLASS.value()))
                 .body("shelterElectricity", equalTo(ElectricityTypeEnumeration.LIGHT.value()))
                 .body("shelterLighting", equalTo(false))
                 .body("shelterCondition", equalTo(ShelterConditionEnumeration.BAD.value()))
@@ -2765,7 +2765,7 @@ public class GraphQLResourceStopPlaceIntegrationTest extends AbstractGraphQLReso
                 .rootPath("data.stopPlace[0].placeEquipments.shelterEquipment[0]")
                 .body(notNullValue())
                 .body("enclosed", equalTo(true))
-                .body("shelterType", equalTo(HslShelterTypeEnumeration.STEEL.value()))
+                .body("shelterType", equalTo(ShelterTypeEnumeration.STEEL.value()))
                 .body("shelterElectricity", equalTo(ElectricityTypeEnumeration.CONTINUOUS.value()))
                 .body("shelterLighting", equalTo(true))
                 .body("shelterCondition", equalTo(ShelterConditionEnumeration.MEDIOCRE.value()))
@@ -2953,7 +2953,7 @@ public class GraphQLResourceStopPlaceIntegrationTest extends AbstractGraphQLReso
         hslAccessibilityProperties.setStopAreaSurroundingsAccessible(false);
         hslAccessibilityProperties.setCurvedStop(true);
         hslAccessibilityProperties.setStopType(HslStopTypeEnumeration.PULL_OUT);
-        hslAccessibilityProperties.setShelterType(ShelterTypeEnumeration.WIDE);
+        hslAccessibilityProperties.setShelterType(ShelterWidthTypeEnumeration.WIDE);
         hslAccessibilityProperties.setGuidanceType(GuidanceTypeEnumeration.BRAILLE);
         hslAccessibilityProperties.setMapType(MapTypeEnumeration.TACTILE);
         hslAccessibilityProperties.setPedestrianCrossingRampType(PedestrianCrossingRampTypeEnumeration.LR);
@@ -2964,7 +2964,7 @@ public class GraphQLResourceStopPlaceIntegrationTest extends AbstractGraphQLReso
     private ShelterEquipment createShelterEquipmentHSl() {
         ShelterEquipment shelterEquipment = new ShelterEquipment();
         shelterEquipment.setEnclosed(true);
-        shelterEquipment.setShelterType(HslShelterTypeEnumeration.GLASS);
+        shelterEquipment.setShelterType(ShelterTypeEnumeration.GLASS);
         shelterEquipment.setShelterElectricity(ElectricityTypeEnumeration.LIGHT);
         shelterEquipment.setShelterLighting(true);
         shelterEquipment.setShelterCondition(ShelterConditionEnumeration.GOOD);
